remove dependency on moto
