+++ Dockerfile.s3
@@ -1,0 +1,146 @@
+# base: Stage which installs necessary runtime dependencies (OS packages, java,...)
+FROM python:3.11.5-slim-buster@sha256:9f35f3a6420693c209c11bba63dcf103d88e47ebe0b205336b5168c122967edf as base
+ARG TARGETARCH
+
+# set workdir
+RUN mkdir -p /opt/code/localstack
+WORKDIR /opt/code/localstack/
+
+# Install runtime OS package dependencies
+RUN --mount=type=cache,target=/var/cache/apt \
+    apt-get update && \
+        # Install dependencies to add additional repos
+        apt-get install -y --no-install-recommends \
+            # Runtime packages (groff-base is necessary for AWS CLI help)
+             ca-certificates curl make openssl
+# do we need the dns server?
+#             ca-certificates curl make openssl iputils-ping iproute2
+# default under
+#             ca-certificates curl gnupg make openssl tar pixz zip unzip groff-base iputils-ping nss-passwords procps iproute2 xz-utils
+
+SHELL [ "/bin/bash", "-c" ]
+
+# create localstack user and filesystem hierarchy, perform some permission fixes
+RUN chmod 777 . && \
+    useradd -ms /bin/bash localstack && \
+    mkdir -p /var/lib/localstack && \
+    chmod -R 777 /var/lib/localstack && \
+    mkdir -p /usr/lib/localstack && \
+    mkdir /tmp/localstack && \
+    chmod -R 777 /tmp/localstack && \
+    touch /tmp/localstack/.marker
+
+# install basic (global) tools to final image
+RUN --mount=type=cache,target=/root/.cache \
+    pip install --no-cache-dir --upgrade virtualenv
+
+# install the entrypoint script
+ADD bin/docker-entrypoint.sh /usr/local/bin/
+# add the shipped hosts file to prevent performance degredation in windows container mode on windows
+# (where hosts file is not mounted) See https://github.com/localstack/localstack/issues/5178
+ADD bin/hosts /etc/hosts
+
+# expose default environment
+# Set edge bind host so localstack can be reached by other containers
+# set library path and default LocalStack hostname
+ENV USER=localstack
+ENV PYTHONUNBUFFERED=1
+
+# Install the latest version of awslocal globally
+# RUN --mount=type=cache,target=/root/.cache \
+#     pip3 install --upgrade awscli awscli-local requests
+
+
+# builder: Stage which installs the dependencies of LocalStack Community
+FROM base as builder
+ARG TARGETARCH
+
+# Install build dependencies to base
+RUN --mount=type=cache,target=/var/cache/apt \
+    apt-get update && \
+        # Install dependencies to add additional repos
+        apt-get install -y gcc
+
+# upgrade python build tools
+RUN --mount=type=cache,target=/root/.cache \
+    (virtualenv .venv && . .venv/bin/activate && pip3 install --upgrade pip wheel setuptools)
+
+# add files necessary to install all dependencies
+ADD Makefile setup.py setup.cfg pyproject.toml ./
+# add the root package init to invalidate docker layers with version bumps
+ADD localstack/__init__.py localstack/
+# add the localstack start scripts (necessary for the installation of the runtime dependencies, i.e. `pip install -e .`)
+ADD bin/localstack bin/localstack.bat bin/localstack-supervisor bin/
+
+# install dependencies to run the LocalStack Pro runtime and save which ones were installed
+RUN --mount=type=cache,target=/root/.cache \
+    make install-s3
+RUN . .venv/bin/activate && pip3 freeze -l > requirements-runtime.txt
+
+# delete the botocore specs for other services (>80mb)
+RUN find .venv/lib/python3.11/site-packages/botocore/data/ -mindepth 1 -maxdepth 1 -type d -not -name s3 -exec rm -rf '{}' \;
+
+
+
+# final stage: Builds upon base stage and copies resources from builder stages
+FROM base
+COPY --from=builder /opt/code/localstack/.venv /opt/code/localstack/.venv
+
+# add project files necessary to install all dependencies
+ADD Makefile setup.py setup.cfg pyproject.toml ./
+# add the localstack start scripts (necessary for the installation of the runtime dependencies, i.e. `pip install -e .`)
+ADD bin/localstack bin/localstack.bat bin/localstack-supervisor bin/
+
+# add the code as late as possible
+ADD localstack/ localstack/
+
+# delete the API specs stubs for other services
+# delete the providers implementations for other services
+RUN find localstack/aws/api/ -mindepth 1 -maxdepth 1 -type d -not -name s3 -exec rm -rf '{}' \; && \
+    find localstack/services/ -mindepth 1 -maxdepth 1 -type d -not -name s3 -exec rm -rf '{}' \;
+
+# Generate the plugin entrypoints
+RUN make entrypoints
+
+# Remove dependecies
+RUN apt-get remove -y make
+
+# remove the other services from the entrypoints
+RUN sed -i '/\[localstack\.aws\.provider\]/,/^$/ { /^\[localstack\.aws\.provider\]/ b; /^$/ b; /localstack.services.providers:s3_v3/!d; }' ./localstack_core.egg-info/entry_points.txt
+
+
+# link the python package installer virtual environments into the localstack venv
+RUN echo /var/lib/localstack/lib/python-packages/lib/python3.11/site-packages > localstack-var-python-packages-venv.pth && \
+    mv localstack-var-python-packages-venv.pth .venv/lib/python*/site-packages/
+RUN echo /usr/lib/localstack/python-packages/lib/python3.11/site-packages > localstack-static-python-packages-venv.pth && \
+    mv localstack-static-python-packages-venv.pth .venv/lib/python*/site-packages/
+
+# expose edge service, external service ports, and debugpy
+EXPOSE 4566 5678
+
+HEALTHCHECK --interval=10s --start-period=15s --retries=5 --timeout=5s CMD ./bin/localstack status services --format=json
+
+# default volume directory
+VOLUME /var/lib/localstack
+
+# mark the image version
+RUN touch /usr/lib/localstack/.s3-version
+
+LABEL authors="LocalStack Contributors"
+LABEL maintainer="LocalStack Team (info@localstack.cloud)"
+LABEL description="LocalStack Docker image"
+
+# Add the build date and git hash at last (changes everytime)
+ARG LOCALSTACK_BUILD_DATE
+ARG LOCALSTACK_BUILD_GIT_HASH
+ARG LOCALSTACK_BUILD_VERSION
+ENV LOCALSTACK_BUILD_DATE=${LOCALSTACK_BUILD_DATE}
+ENV LOCALSTACK_BUILD_GIT_HASH=${LOCALSTACK_BUILD_GIT_HASH}
+ENV LOCALSTACK_BUILD_VERSION=${LOCALSTACK_BUILD_VERSION}
+ENV EAGER_SERVICE_LOADING=1
+ENV SERVICES=s3
+# TODO: do we need DNS for the S3 image?
+ENV DNS_ADDRESS=false
+
+# define command at startup
+ENTRYPOINT ["docker-entrypoint.sh"]
