+++ localstack/services/s3/notifications.py
@@ -7,8 +7,7 @@ from typing import Dict, List, Optional, Tuple, TypedDict, Union
 from urllib.parse import quote
 
 from botocore.exceptions import ClientError
-from moto.s3.models import FakeBucket, FakeDeleteMarker, FakeKey
-
+from localstack import config
 from localstack.aws.api import RequestContext
 from localstack.aws.api.events import PutEventsRequestEntry
 from localstack.aws.api.lambda_ import InvocationType
@@ -33,12 +32,7 @@ from localstack.aws.api.s3 import (
     TopicConfiguration,
 )
 from localstack.aws.connect import connect_to
-from localstack.services.s3.models import get_moto_s3_backend
-from localstack.services.s3.utils import (
-    _create_invalid_argument_exc,
-    get_bucket_from_moto,
-    get_key_from_moto_bucket,
-)
+from localstack.services.s3.utils import _create_invalid_argument_exc
 from localstack.services.s3.v3.models import S3Bucket, S3DeleteMarker, S3Object
 from localstack.utils.aws import arns
 from localstack.utils.aws.arns import parse_arn, s3_bucket_arn
@@ -46,6 +40,14 @@ from localstack.utils.aws.client_types import ServicePrincipal
 from localstack.utils.bootstrap import is_api_enabled
 from localstack.utils.strings import short_uid
 from localstack.utils.time import parse_timestamp, timestamp_millis
+
+if config.LEGACY_V2_S3_PROVIDER:
+    from moto.s3.models import FakeBucket, FakeDeleteMarker, FakeKey
+    from localstack.services.s3.utils_moto import (
+        get_bucket_from_moto,
+        get_key_from_moto_bucket,
+    )
+    from localstack.services.s3.models import get_moto_s3_backend
 
 LOG = logging.getLogger(__name__)
 
@@ -327,9 +329,7 @@ class BaseNotifier:
         - validate the arn pattern
         - validating the Rule names (and normalizing to capitalized)
         - check if the filter value is not empty
-        :param configuration: a configuration for a notifier, like QueueConfiguration for SQS or
-        TopicConfiguration for SNS
-        :param skip_destination_validation: allow skipping the validation of the target
+        :param verification_ctx: the verification context containing necessary data for validation
         :raises InvalidArgument if the rule is not valid, infos in ArgumentName and ArgumentValue members
         :return:
         """
@@ -820,10 +820,10 @@ class NotificationDispatcher:
             configurations = (
                 configurations if isinstance(configurations, list) else [configurations]
             )
-            for config in configurations:
-                if notifier.should_notify(ctx, config):  # we check before sending it to the thread
+            for configuration in configurations:
+                if notifier.should_notify(ctx, configuration):  # we check before sending it to the thread
                     LOG.debug("Submitting task to the executor for notifier %s", notifier)
-                    self.executor.submit(notifier.notify, ctx, config)
+                    self.executor.submit(notifier.notify, ctx, configuration)
 
     def verify_configuration(
         self,
