+++ setup.cfg
@@ -86,10 +86,28 @@ runtime =
     Quart>=0.19.2
     readerwriterlock>=1.0.7
     requests-aws4auth>=1.0
-    # explicitly set urllib3 to force its usage / ensure compatibility
-    urllib3>=2.0.7
     Werkzeug>=3.0.0
     xmltodict>=0.13.0
+
+# required to actually run localstack on the host
+s3 =
+    awscrt>=0.13.14
+    boto3>=1.26.121
+    botocore>=1.31.2
+    cbor2>=5.2.0
+    dnspython>=1.16.0
+    docker==6.0.1
+    jsonpatch>=1.24
+    hypercorn>=0.14.4
+    localstack-client>=2.0
+    pyopenssl>=23.0.0
+    readerwriterlock>=1.0.7
+    requests-aws4auth>=1.0
+    Werkzeug>=3.0.0
+    xmltodict>=0.13.0
+    # TODO: remove those 2 dependencies
+    flask>=3.0.0
+    Quart>=0.19.2
 
 # @deprecated - use extra 'runtime' instead.
 full =
